refactor(utils/react-hooks): Rename `useDualTriggerEvent` to `useDoubleTrigger`
